Fix sign conversions and add the warning check to the CMake (#800)
IoT.js-DCO-1.0-Signed-off-by: László Langó llango.u-szeged@partner.samsung.com

+++ src/iotjs.c
@@ -173,7 +173,8 @@ int iotjs_entry(int argc, char** argv) {
   iotjs_environment_t* env = (iotjs_environment_t*)iotjs_environment_get();
 
   // Parse command line arguments.
-  if (!iotjs_environment_parse_command_line_arguments(env, argc, argv)) {
+  if (!iotjs_environment_parse_command_line_arguments(env, (uint32_t)argc,
+                                                      argv)) {
     DLOG("iotjs_environment_parse_command_line_arguments failed");
     return 1;
   }
